requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Unidecode==0.04.16
Werkzeug==0.9.6
argparse==1.2.1
gunicorn==19.1.1
itsdangerous==0.24
python-slugify==0.1.0
wsgiref==0.1.2
